refactor(code_gen): extract BinaryOp into binop.rs
